Better namespacing and separate type check validator
